some small modifications that need implementations

+++ src/main/java/proiect_fic/smarttrafficmanagementsystem/models/Memorie.java
@@ -11,9 +11,9 @@ public class Memorie {
     private static int KINIT = 1;
 
     private int[] debite = new int[4];
-    private int[] coef = new int[4];
+    private float[] coef = new float[4];
 
-    public Memorie(int[] debite, int[] coef) {
+    public Memorie(int[] debite, float[] coef) {
         this.debite = debite;
         this.coef = coef;
     }
@@ -43,10 +43,11 @@ public class Memorie {
         return this.debite;
     }
 
-    public int[] getCoef() {
+    public float[] getCoef() {
         return this.coef;
     }
-    public int getCoefSemafor(int index) {
+
+    public float getCoefSemafor(int index) {
         return coef[index];
     }
 
@@ -59,17 +60,17 @@ public class Memorie {
     public void setDebite(int[] debite) {
         this.debite = debite;
     }
-    public void setDebitSemafor(int index, int d)
-    {
-        this.debite[index]=d;
+
+    public void setDebitSemafor(int index, int d) {
+        this.debite[index] = d;
     }
-    public void setCoef(int[] coef) {
+
+    public void setCoef(float[] coef) {
         this.coef = coef;
     }
 
-    public void setCoefSemafor(int index, int c)
-    {
-        this.coef[index]=c;
+    public void setCoefSemafor(int index, float c) {
+        this.coef[index] = c;
     }
 
 }
